requirements/base.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic-settings==2.0.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
redis==5.0.1
celery==5.3.4
httpx==0.25.1
python-multipart==0.0.6
structlog==24.1.0
prometheus-client==0.19.0
tiktoken==0.5.1
openai==1.3.0
anthropic==0.7.7
pinecone-client==3.0.0
chromadb==0.4.22
sentence-transformers==2.2.2
pypdf==3.17.4
python-magic==0.4.27
apscheduler==3.10.4
